interface/requirements.txt
PyQt6>=6.4
pyqtgraph>=0.13
pyserial>=3.5
numpy>=1.24
pandas>=2.0
